# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rag_chatbot_backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/documents/zip-sessions": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create Zip Markdown Session",
        "description": "Phase 1: store ZIP and return a flat sorted manifest (root + nested paths). No embeddings.",
        "operationId": "create_zip_markdown_session_api_v1_documents_zip_sessions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_zip_markdown_session_api_v1_documents_zip_sessions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZipSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/documents/zip-sessions/{session_id}/ingest": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Ingest Zip Session Batch",
        "description": "Phase 2: chunk + OpenAI embeddings + Qdrant for one path batch (configurable skip/limit).",
        "operationId": "ingest_zip_session_batch_api_v1_documents_zip_sessions__session_id__ingest_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZipBulkIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "description": "Optional JSON body; omit or send `{}` to use session defaults for skip/limit slicing.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZipIngestBatchRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/zip-sessions/{session_id}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Zip Markdown Session",
        "operationId": "delete_zip_markdown_session_api_v1_documents_zip_sessions__session_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "type": "array",
                  "title": "Response List Documents Api V1 Documents Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/websites": {
      "get": {
        "tags": [
          "websites"
        ],
        "summary": "List Websites",
        "operationId": "list_websites_api_v1_websites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebsiteResponse"
                  },
                  "type": "array",
                  "title": "Response List Websites Api V1 Websites Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "websites"
        ],
        "summary": "Create Website",
        "operationId": "create_website_api_v1_websites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/websites/{website_id}": {
      "get": {
        "tags": [
          "websites"
        ],
        "summary": "Get Website",
        "operationId": "get_website_api_v1_websites__website_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "websites"
        ],
        "summary": "Update Website",
        "operationId": "update_website_api_v1_websites__website_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/websites/{website_id}/crawl": {
      "post": {
        "tags": [
          "websites"
        ],
        "summary": "Crawl Website Now",
        "operationId": "crawl_website_now_api_v1_websites__website_id__crawl_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/websites/{website_id}/crawl-jobs": {
      "get": {
        "tags": [
          "websites"
        ],
        "summary": "List Website Crawl Jobs",
        "operationId": "list_website_crawl_jobs_api_v1_websites__website_id__crawl_jobs_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrawlJobResponse"
                  },
                  "title": "Response List Website Crawl Jobs Api V1 Websites  Website Id  Crawl Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_zip_markdown_session_api_v1_documents_zip_sessions_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_zip_markdown_session_api_v1_documents_zip_sessions_post"
      },
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
...
      "CrawlFrequency": {
        "type": "string",
        "enum": [
          "never",
          "only_once",
          "12h",
          "1d"
        ],
        "title": "CrawlFrequency"
      },
      "CrawlJobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "website_id": {
            "type": "string",
            "title": "Website Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "status": {
            "$ref": "#/components/schemas/CrawlJobStatus"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "pages_crawled": {
            "type": "integer",
            "title": "Pages Crawled"
          },
          "chunks_indexed": {
            "type": "integer",
            "title": "Chunks Indexed"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "website_id",
          "user_id",
          "status",
          "started_at",
          "finished_at",
          "pages_crawled",
          "chunks_indexed",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "CrawlJobResponse"
      },
      "CrawlJobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "CrawlJobStatus"
      },
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "vector_count": {
            "type": "integer",
            "title": "Vector Count"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "zip_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip Session Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "original_filename",
          "document_type",
          "content_type",
          "size_bytes",
          "status",
          "chunk_count",
          "vector_count",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "uploaded",
          "processing",
          "indexed",
          "failed"
        ],
        "title": "DocumentStatus"
      },
      "DocumentType": {
        "type": "string",
        "enum": [
          "pdf",
          "markdown",
          "markdown_zip"
        ],
        "title": "DocumentType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WebsiteCreate": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "frequency": {
            "$ref": "#/components/schemas/CrawlFrequency",
            "default": "never"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "WebsiteCreate"
      },
      "WebsiteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "frequency": {
            "$ref": "#/components/schemas/CrawlFrequency"
          },
          "status": {
            "$ref": "#/components/schemas/WebsiteStatus"
          },
          "last_crawled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Crawled At"
          },
          "next_crawl_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Crawl At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "url",
          "name",
          "frequency",
          "status",
          "last_crawled_at",
          "next_crawl_at",
          "last_error",
          "created_at",
          "updated_at"
        ],
        "title": "WebsiteResponse"
      },
      "WebsiteStatus": {
        "type": "string",
        "enum": [
          "active",
          "paused"
        ],
        "title": "WebsiteStatus"
      },
      "WebsiteUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "frequency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CrawlFrequency"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WebsiteStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "WebsiteUpdate"
      },
      "ZipBulkIngestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "vector_count": {
            "type": "integer",
            "title": "Vector Count"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "zip_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip Session Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "files_indexed": {
            "type": "integer",
            "title": "Files Indexed",
            "default": 0
          },
          "files_skipped": {
            "type": "integer",
            "title": "Files Skipped",
            "default": 0
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "markdown_files_total": {
            "type": "integer",
            "title": "Markdown Files Total",
            "default": 0
          },
          "markdown_skip": {
            "type": "integer",
            "title": "Markdown Skip",
            "default": 0
          },
          "markdown_limit": {
            "type": "integer",
            "title": "Markdown Limit",
            "default": 0
          },
          "has_more_markdown_files": {
            "type": "boolean",
            "title": "Has More Markdown Files",
            "default": false
          },
          "next_markdown_skip": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Markdown Skip"
          },
          "indexed_document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Indexed Document Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "original_filename",
          "document_type",
          "content_type",
          "size_bytes",
          "status",
          "chunk_count",
          "vector_count",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "ZipBulkIngestResponse",
        "description": "Response for one ingest batch (one Mongo row + Qdrant points per inner Markdown file)."
      },
...
      "ZipSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "zip_size_bytes": {
            "type": "integer",
            "title": "Zip Size Bytes"
          },
          "markdown_files": {
            "items": {
              "$ref": "#/components/schemas/ZipMarkdownFileEntry"
            },
            "type": "array",
            "title": "Markdown Files"
          },
          "manifest_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Manifest Warnings"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "default_path_batch": {
            "type": "integer",
            "title": "Default Path Batch"
          },
          "max_path_batch": {
            "type": "integer",
            "title": "Max Path Batch"
          },
          "max_path_indices_per_request": {
            "type": "integer",
            "title": "Max Path Indices Per Request"
          },
          "openai_embedding_text_batch": {
            "type": "integer",
            "title": "Openai Embedding Text Batch"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id",
          "original_filename",
          "zip_size_bytes",
          "markdown_files",
          "expires_at"
        ],
        "title": "ZipSessionResponse",
        "description": "Phase 1: flat manifest only (no embeddings)."
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}